+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/IrMetroContext.kt
@@ -7,6 +7,9 @@ import dev.zacsweers.metro.compiler.MetroLogger
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.compat.CompatContext
 import dev.zacsweers.metro.compiler.exitProcessing
+import dev.zacsweers.metro.compiler.ir.cache.IrCache
+import dev.zacsweers.metro.compiler.ir.cache.IrCachesFactory
+import dev.zacsweers.metro.compiler.ir.cache.IrThreadUnsafeCachesFactory
 import dev.zacsweers.metro.compiler.symbols.Symbols
 import dev.zacsweers.metro.compiler.tracing.Tracer
 import dev.zacsweers.metro.compiler.tracing.tracer
@@ -27,13 +30,10 @@ import org.jetbrains.kotlin.incremental.components.Position
 import org.jetbrains.kotlin.incremental.components.ScopeKind
 import org.jetbrains.kotlin.ir.IrElement
 import org.jetbrains.kotlin.ir.declarations.IrClass
-import org.jetbrains.kotlin.ir.types.IrType
 import org.jetbrains.kotlin.ir.types.IrTypeSystemContext
 import org.jetbrains.kotlin.ir.types.IrTypeSystemContextImpl
-import org.jetbrains.kotlin.ir.util.TypeRemapper
 import org.jetbrains.kotlin.ir.util.dumpKotlinLike
 import org.jetbrains.kotlin.ir.util.parentDeclarationsWithSelf
-import org.jetbrains.kotlin.name.ClassId
 
 internal interface IrMetroContext : IrPluginContext, CompatContext {
   val metroContext
@@ -60,7 +60,16 @@ internal interface IrMetroContext : IrPluginContext, CompatContext {
   val lookupFile: Path?
   val expectActualFile: Path?
 
-  val typeRemapperCache: MutableMap<Pair<ClassId, IrType>, TypeRemapper>
+  /**
+   * Generic caching machinery. Add new caches as extension functions that encapsulate the [key] and
+   * types.
+   *
+   * @param key A unique string identifier for this cache
+   */
+  fun <K : Any, V : Any, C> getOrCreateIrCache(
+    key: Any,
+    createCache: (IrCachesFactory) -> IrCache<K, V, C>,
+  ): IrCache<K, V, C>
 
   fun onErrorReported()
 
@@ -245,8 +254,16 @@ internal interface IrMetroContext : IrPluginContext, CompatContext {
         }
       }
 
-      override val typeRemapperCache: MutableMap<Pair<ClassId, IrType>, TypeRemapper> =
-        mutableMapOf()
+      private val genericCaches: HashMap<Any, IrCache<*, *, *>> = HashMap()
+
+      override fun <K : Any, V : Any, C> getOrCreateIrCache(
+        key: Any,
+        createCache: (IrCachesFactory) -> IrCache<K, V, C>,
+      ): IrCache<K, V, C> {
+        @Suppress("UNCHECKED_CAST")
+        return genericCaches.getOrPut(key) { createCache(IrThreadUnsafeCachesFactory) }
+          as IrCache<K, V, C>
+      }
     }
   }
 }